[COOK-2465] Add compile and vim settings recipes.
This commit adds two new recipes and attributes to the default recipe that
gives the user the option to compile vim from source as well as add a very
robust vim settings.

Both the compile and settings recipe are highly configurable and allow the
user to override almost all the default settings.

This commit does not change the default behavior of the original vim cookbook
unless it's configured to.

Attemps have been made to only use Chef resources in the new recipes so as to
keep this cookbook indempotent.

Conflicts:
	recipes/default.rb