Don't break subsequent chef runs in the event of a failure
Previously if we downloaded a file and then failed to compile it, then vim would never install unless you SSH'd to the node and removed the tarball.  Now if the app hasn't been installed we'll always try to install it.  This goes hand in hand with the previous fix for make clean so we can try to compile again after a failed compile.